project/web_e_business/requirements.txt
Automat==20.2.0
Django==2.2.6
Flask==1.1.2
Flask-WTF==0.14.3
JPype1==1.1.2
Jinja2==2.11.1
Markdown==3.3.3
MarkupPy==1.14
MarkupSafe==1.1.1
Pillow==6.2.0
protego==0.1.16
PyDispatcher==2.0.5
PyHamcrest==2.0.2
PyMySQL==0.9.3
PyYAML==5.1.2
Pygments==2.6.1
QtPy==1.9.0
Scrapy==2.4.1
Send2Trash==1.5.0
Twisted==20.3.0
WTForms==2.3.3
Werkzeug==1.0.1
asgiref==3.2.7
attrs==19.2.0
backcall==0.1.0
beautifulsoup4==4.9.0
bleach==3.1.3
bs4==0.0.1
captcha==0.3
certifi==2019.9.11
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.3
constantly==15.1.0
cryptography==3.2.1
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
django-cors-headers==3.2.1
django-crispy-forms==1.8.0
django-filter==2.4.0
django-formtools==2.1
django-import-export==1.2.0
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-recaptcha==2.0.6
django-simple-captcha==0.5.12
djangorestframework==3.12.1
entrypoints==0.3
et-xmlfile==1.0.1
future==0.18.1
httplib2==0.14.0
hyperlink==20.0.1
idna==2.8
importlib-metadata==1.5.0
incremental==17.5.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itemadapter==0.2.0
itemloaders==1.0.4
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.16.0
jmespath==0.10.0
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.0
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.1.0
lxml==4.5.0
matplotlib==3.2.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.2
odfpy==1.4.0
openpyxl==3.0.0
pandas==1.0.3
pip==20.2.4
prettytable==2.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.4
pyOpenssL==20.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyecharts==1.9.0
pyparsing==2.4.6
pyrsistent==0.16.0
python-dateutil==2.8.1
python-gitlab==1.13.0
pytz==2019.3
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.0
qtconsole==4.7.1
queuelib==1.5.0
redis==3.5.3
requests==2.22.0
scikit-learn==0.22.2.post1